add component for joining columns
