+++ app/templates/components/select-column.hbs
@@ -1,5 +1,5 @@
 <div class={{if (eq properties.function "join") "hidden"}}>
-  {{#ui-dropdown class="selection fluid" selected=properties.fields.firstObject onChange=(action "sendChooseField" field) as |execute mapper|}}
+  {{#ui-dropdown class="selection fluid" selected=properties.fields.firstObject onChange=(action "chooseField" field) as |execute mapper|}}
     <div class="default text">
       Source column headings
     </div>
